1st stab at unattended installation.
* Introduce answers file for rake to read from, instead of STDIN,
when it needs to know whether to install a feature or not.
* Add a rake task to generate a default answers file that says yes
to all features, with a comment at the top regarding answer syntax.
* Branch execution so that if an answers file exists, we consider
assume unattended installation.

TODO: Figure out how to do something similar for the overwrite/
backup prompts.  Ideas welcome.  Simplest is just to backup all
files, I guess, and let the user diff them if he wants.  Rationale
being that for an unattended install, there are either
a) probably no user customisations in the first place, since it'll
be a from bare-metal provisioning scenario
b) or specific pre-customised files to be dropped in over the top
anyway for site-specific consistency.
... at any rate, that's my scenario ;-)