Add solarized themes for iTerm2
Both solarized dark and light are now included in the subdir
iTerm2. They are copies from the original repo because it doesn't
seem worth to clone a ~10Mb repo just to keep in sync 2 files,
considering that solarized hasn't changed much in the past years.

The rakefile has been updated to trigger the automatic installation
of these two themes only if OS == OSX. The user will still have
to set the themes for his/her current profile manually (but that's
explained at the end of the installation process).