Fix a bug that was causing .vim.after to be removed after update
rake install/update was executing git clean with the -x flag that
force the removal of ignored files. Since .vim.after was listed
inside .gitignore, this was causing the file to be removed on
each update.
I don't think we need to remove ignored files since are all
temporary or user-specific configuration files that should
persist between updates.

Fixes #225