Quick Buffer Navigationw with BuffExplorer
BuffExplorer provides a complete Buffer list that is accessible (<Tab>)
and easy on the brain by providing a complete vertical list of buffers.

This way you don't have to remember the name of the file you're working
on (like with LustyJuggler). If you remember the name of the file, you
can use `/term<CR>` to open the first matching file.