Qargs & Qargsdo for executing actions on quickfix lists
The immediate usage for this is for replacement in conjunction with
`Ag`. This solution is much quicker than using `:Gsearch` and
`:Greplace`, not only in the amount of steps, but also in the
replacement performance when the amount of files is high.

Usually `Gsearch` will freeze and `Qargs | Argdo` will execute
successfully in less time.

README: https://github.com/MarioRicalde/vim-qargs