BuffExplorer Implementation BuffExplorer provides a complete Buffer list that is accessible (`<Tab>`) and easy on the brain by providing a complete vertical list of buffers.
This way you don't have to remember the name of the file you're working
on (like with LustyJuggler). If you remember the name of the file, you
can use `/term<CR>` to open the first matching file.

I've modified the syntax file of the plugin to make it comply with the
"it should look good" rule of YADR.

Hopefully BuffExplorer will proof superior to LustyJuggler and
eventually replace it.

Screenshot of how it looks

![](https://dl.dropbox.com/s/2rpci91mqdt97ge/2013-06-03_at_5.34.17_AM.png)

I noticed `,b` but it is slow in two ways:

1. It's slow by loading since it depends on the `<leader>` key delay.
2. You need to remember the name of the file, and if you want to scroll
you move your hands away from home keys.

With BufExplorer you will begin doing `<Tab>` `j` or `k`, it's superior.
Also it's close to `CTRL` which is your `<ESC>`